refactor(channels): converge operational↔phraseAck spoken rules to SSOT
Share colloquial/one-line policy atoms across ack and operational paths, keep intentional differences (conditional naming, 执行环境 vs 沙箱, exclusive bans), split voice/phrase/operational builders out of live.go, and lock the alignment with a regression test.
